requirements.txt
matplotlib>=3.4.2
numpy>=1.19.3
tqdm
antlr4-python3-runtime~=4.13.1
py_trees~=2.2.3
shortuuid~=1.0.11
autopep8~=2.0.4
openai~=1.34.0
sympy~=1.12
gymnasium~=0.29.1
minigrid~=2.3.1
requests~=2.32.3
opencv-python~=4.10.0.84
ipdb~=0.13.13
networkx~=3.3
tabulate~=0.9.0
pandas~=2.2.2
hydra-core~=1.3.2
latextable~=1.0.1
tdw~=1.13.0.0